+++ cognite/neat/rules/importers/_owl2rules/_owl2metadata.py
